Revert "feat + refactor: add string spec component and separate TestNo and Te…"
