+++ src/reducers/testgen.js
@@ -71,42 +71,3 @@ export function verticalArray(state={checkedVerticalArray: false}, action){
   return state
 }
 
-export function lowercase(state={lowercaseChecked:true}, action){
-  switch(action.type){
-    case 'CHECK_LOWERCASE':
-      return {lowercaseChecked: action.checked}
-  }
-  return state
-}
-
-export function uppercase(state={uppercaseChecked:false}, action){
-  switch(action.type){
-    case 'CHECK_UPPERCASE':
-      return {uppercaseChecked: action.checked}
-  }
-  return state
-}
-
-export function digits(state={digitsChecked:false}, action){
-  switch(action.type){
-    case 'CHECK_DIGITS':
-      return {digitsChecked: action.checked}
-  }
-  return state
-}
-
-export function symbols(state={symbolsChecked:false}, action){
-  switch(action.type){
-    case 'CHECK_SYMBOLS':
-      return {symbolsChecked: action.checked}
-  }
-  return state
-}
-
-export function stringLength(state={stringLength: 1}, action){
-  switch(action.type){
-    case 'CHANGE_STRING_LENGTH':
-      return { stringLength: action.value }
-  }
-  return state
-}
